+++ src/core/can.h
@@ -37,7 +37,7 @@ status_t can_print_channel_help(core_t* core);
 void can_print_error(uint32 can_id, const char* reason, disp_mode_t disp_mode);
 void can_set_baud_rate(uint8 baud_rate_index, core_t* core);
 void can_set_channel(uint32 channel, core_t* core);
-void limit_node_id(uint8* node_id);
+void limit_node_id(uint32* node_id);
 bool is_can_initialised(core_t* core);
 
 #endif /* CAN_H */
